- 修改各圖表Render類別結構 - 完善圖表example - 添加圖表的線圖閃耀動畫(但因為會不斷重建圖表, 先暫時隱藏此功能)

+++ lib/ui/widget/k_line_chart/model/k_line_data/k_line_data.dart
@@ -21,7 +21,7 @@ class KLineData {
   final double close;
 
   /// 成交量(手數)
-  final int volume;
+  final double volume;
 
   /// 成交金額
   final double amount;
